+++ CHANGELOG.md
@@ -1,5 +1,12 @@
-# Changelog
+## 1.1.0
+
+* Add `start` and `end` parameters to the `addAll()` and `insertAll()` methods
+  for the typed data buffer classes. These allow efficient concatenation of
+  slices of existing typed data.
+
+* Make `addAll()` for typed data buffer classes more efficient for lists,
+  especially typed data lists.
 
 ## 1.0.0
 
-- ChangeLog starts here
+* ChangeLog starts here